Fortify FPR converter fixes (#2014)
* FortifyFpr converter improvements.

* Fix typo in file.

* Code changes to convrter

* Add DSP ingestion visitor.

* Files to demonstrate Fortify DSP progress.

* Update script to auto-gen drive letter based on script path.

* Update replacement logic.

* Update tests based on foritfy fopr converter and page command improvements.

* Remove test files for now.

* Delete test files.:

+++ src/Sarif.Multitool/CommandBase.cs
@@ -16,7 +16,7 @@ namespace Microsoft.CodeAnalysis.Sarif.Multitool
         internal const int SUCCESS = 0;
         internal const int FAILURE = 1;
 
-        protected static bool ValidateNonNegativeCommandLineOption<T>(int optionValue, string optionName)
+        protected static bool ValidateNonNegativeCommandLineOption<T>(long optionValue, string optionName)
         {
             bool valid = true;
